Finish fixing up from the move
